% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwellCount.R
\name{dwellCount}
\alias{dwellCount}
\title{This function obtains dwelling times for states from a list.}
\usage{
dwellCount(x)
}
\arguments{
\item{x}{List with states}
}
\description{
This function obtains dwelling times for states from a list.
}
