% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_dwell.R
\name{nest_dwell}
\alias{nest_dwell}
\title{This function estimates the dwelling time or continuous occupancy of
a series of states.'}
\usage{
nest_dwell(tbl, vars, foVar, sortBy)
}
\arguments{
\item{tbl}{A tibble/dataframe featuring a list of states in time
- and covariates}

\item{vars}{Variables of interest in the tibble/dataframe}

\item{foVar}{States variable name to obtain Dwelling time}

\item{sortBy}{Time index variable}
}
\description{
This function estimates the dwelling time or continuous occupancy of
a series of states.'
}
